refactor : Entity 클래스 Model.Entity 패키지로 이동
