+++ src/main/java/gift/Repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package gift.Repository;
 
-import gift.Model.Category;
+import gift.Model.Entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
