+++ src/main/java/gift/Repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package gift.Repository;
 
-import gift.Model.Member;
+import gift.Model.Entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
