+++ src/main/java/gift/Repository/OptionRepository.java
@@ -1,9 +1,9 @@
 package gift.Repository;
 
-import gift.Model.Option;
-import gift.Model.Product;
+import gift.Model.Entity.Option;
+import gift.Model.Entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
+
 import java.util.List;
 
 public interface OptionRepository extends JpaRepository<Option, Long> {
