+++ src/main/java/gift/Repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package gift.Repository;
 
-import gift.Model.Product;
+import gift.Model.Entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product,Long> {
