+++ src/main/java/gift/Service/ProductService.java
@@ -4,7 +4,9 @@ import gift.DTO.RequestProductDTO;
 import gift.DTO.RequestProductPostDTO;
 import gift.Exception.CategoryNotFoundException;
 import gift.Exception.ProductNotFoundException;
-import gift.Model.*;
+import gift.Model.Entity.Category;
+import gift.Model.Entity.Option;
+import gift.Model.Entity.Product;
 import gift.Repository.CategoryRepository;
 import gift.Repository.OptionRepository;
 import gift.Repository.ProductRepository;
