+++ src/main/java/gift/Service/WishService.java
@@ -4,7 +4,9 @@ import gift.DTO.RequestWishDTO;
 import gift.DTO.ResponseWishDTO;
 import gift.Exception.ProductNotFoundException;
 import gift.Exception.WishNotFoundException;
-import gift.Model.*;
+import gift.Model.Entity.Member;
+import gift.Model.Entity.Product;
+import gift.Model.Entity.Wish;
 import gift.Repository.ProductRepository;
 import gift.Repository.WishRepository;
 import org.springframework.data.domain.Page;
