+++ src/test/java/gift/EntityTest/ProductTest.java
@@ -1,7 +1,7 @@
 package gift.EntityTest;
 
-import gift.Model.Category;
-import gift.Model.Product;
+import gift.Model.Entity.Category;
+import gift.Model.Entity.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
